Revert "Feature/holiday"

+++ src/main/java/Chopsticks/HairHaeJoBackend/service/DesignerHolidayService.java
@@ -31,11 +31,18 @@ public class DesignerHolidayService {
     private final DesignerHolidayRepository designerHolidayRepository;
     private final UserRepository userRepository;
     public void post(HolidayDto holidayDto) throws IOException {
-        DesignerHoliday holiday =findHoliday();
-        if(holiday!=null) addeach(holidayDto,holiday);
+
+
+        long currentId= SecurityUtil.getCurrentMemberId();
+        User user=userRepository.findById(currentId)
+                .orElseThrow(() -> new RuntimeException("로그인 상태가 아닙니다"));
+
+        if(user.getRole() != Role.ROLE_DESIGNER) throw new RuntimeException("헤어디자이너만 접근 가능합니다");
+        DesignerHoliday holiday =designerHolidayRepository.findBydesignerId(currentId);
+        if(holiday!=null) addeach(holidayDto);
         else {
             try {
-                designerHolidayRepository.save(holidayDto.toholiday(SecurityUtil.getCurrentMemberId()));
+                designerHolidayRepository.save(holidayDto.toholiday(currentId));
             } catch (Exception e) {
                 throw new RuntimeException("휴일 작성을 실패했습니다");
             }
@@ -43,9 +50,63 @@ public class DesignerHolidayService {
 
     }
 
-    public void deleteeach(HolidayDto holidayDto) throws IOException {
-        DesignerHoliday holiday =findHoliday();
+    public void change(HolidayDto holidayDto) throws IOException {
+
+
+        long currentId= SecurityUtil.getCurrentMemberId();
+        User user=userRepository.findById(currentId)
+                .orElseThrow(() -> new RuntimeException("로그인 상태가 아닙니다"));
+
+        if(user.getRole() != Role.ROLE_DESIGNER) throw new RuntimeException("헤어디자이너만 접근 가능합니다");
         try {
+
+            DesignerHoliday holiday =designerHolidayRepository.findBydesignerId(currentId);
+            if(holiday==null) throw new RuntimeException();
+            else {
+                holiday.setDesignerHoliday(holidayDto.getHoliday());
+                designerHolidayRepository.save(holiday);
+            }
+
+        }
+        catch(Exception e) {
+            throw new RuntimeException("휴일 작성을 실패했습니다");
+        }
+
+    }
+    public void delete() throws IOException {
+
+
+        long currentId= SecurityUtil.getCurrentMemberId();
+        User user=userRepository.findById(currentId)
+                .orElseThrow(() -> new RuntimeException("로그인 상태가 아닙니다"));
+
+        if(user.getRole() != Role.ROLE_DESIGNER) throw new RuntimeException("헤어디자이너만 접근 가능합니다");
+        try {
+
+            DesignerHoliday holiday =designerHolidayRepository.findBydesignerId(currentId);
+            if(holiday==null) throw new RuntimeException();
+            else {
+
+                designerHolidayRepository.delete(holiday);
+            }
+
+        }
+        catch(Exception e) {
+            throw new RuntimeException("휴일 삭제를 실패했습니다");
+        }
+
+    }
+    public void deleteeach(HolidayDto holidayDto) throws IOException {
+
+
+        long currentId= SecurityUtil.getCurrentMemberId();
+        User user=userRepository.findById(currentId)
+                .orElseThrow(() -> new RuntimeException("로그인 상태가 아닙니다"));
+        DesignerHoliday holiday;
+        if(user.getRole() != Role.ROLE_DESIGNER) throw new RuntimeException("헤어디자이너만 접근 가능합니다");
+        try {
+
+            holiday =designerHolidayRepository.findBydesignerId(currentId);
             if(holiday==null) throw new RuntimeException();
             String nowHoliday=DeleteHoliday(Arrays.asList(holidayDto.getHoliday().split(",")),Arrays.asList(holiday.getDesignerHoliday().split(",")));
             holiday.setDesignerHoliday(nowHoliday);
@@ -55,33 +116,65 @@ public class DesignerHolidayService {
             throw new RuntimeException("휴일 삭제를 실패했습니다");
         }
 
-    }
-    public void addeach(HolidayDto holidayDto,DesignerHoliday holiday) throws IOException {
 
+
+
+
+
+    }
+    public void addeach(HolidayDto holidayDto) throws IOException {
+
+
+        long currentId= SecurityUtil.getCurrentMemberId();
+        User user=userRepository.findById(currentId)
+                .orElseThrow(() -> new RuntimeException("로그인 상태가 아닙니다"));
+        DesignerHoliday holiday;
+
+        if(user.getRole() != Role.ROLE_DESIGNER) throw new RuntimeException("헤어디자이너만 접근 가능합니다");
         try {
+
+            holiday =designerHolidayRepository.findBydesignerId(currentId);
+            if(holiday==null) throw new RuntimeException();
             String nowHoliday=AddHoliday(Arrays.asList(holidayDto.getHoliday().split(",")),Arrays.asList(holiday.getDesignerHoliday().split(",")));
             holiday.setDesignerHoliday(nowHoliday);
+            System.out.println(nowHoliday);
             designerHolidayRepository.save(holiday);
+
         }
         catch(Exception e) {
             throw new RuntimeException("휴일 추가를 실패했습니다");
         }
 
+
+
+
+
+
     }
 
     public  ArrayList<HolidayDto>  view(long holidayDto) throws IOException {
-        DesignerHoliday holiday =findHoliday();
-        ArrayList<HolidayDto> returndata;
-        try {
 
+
+        long currentId= SecurityUtil.getCurrentMemberId();
+        User user=userRepository.findById(currentId)
+                .orElseThrow(() -> new RuntimeException("로그인 상태가 아닙니다"));
+
+        if(user.getRole() != Role.ROLE_DESIGNER) throw new RuntimeException("헤어디자이너만 접근 가능합니다");
+
+        DesignerHoliday holiday=null;
+        try {
+            holiday =designerHolidayRepository.findBydesignerId(currentId);
             if(holiday==null) throw new RuntimeException();
-            returndata=getHoliday(holiday.getDesignerHoliday().split(","));
+
+
+
+
         }
         catch(Exception e) {
             throw new RuntimeException("휴일 확인을 실패했습니다");
         }
-
-        return returndata;
+        ArrayList<HolidayDto> data=getHoliday(holiday.getDesignerHoliday().split(","));
+        return data;
 
 
     }
@@ -96,22 +189,34 @@ public class DesignerHolidayService {
 
     private String DeleteHoliday(List<String> deleteholiday,List<String> holiday) {
 
+
+        Iterator<String> it = holiday.iterator();
         List<String> newHoliday= new ArrayList<>();
-        for (String nowHoliday:holiday) {
+        while (it.hasNext()) {
+            String item = it.next();
             boolean checkdelete=false;
             for(String Delete:deleteholiday) {
-                if(nowHoliday.equals(Delete))  {
+                if(item.equals(Delete))  {
                     checkdelete=true;
                 }
             }
-            if(!checkdelete)  newHoliday.add(nowHoliday);
+            if(!checkdelete)  newHoliday.add(item);
+
+
         }
-        return String.join(",", newHoliday.toArray(new String[0]));
+
+
+        String[] returnlist= newHoliday.toArray(new String[0]);
+        return String.join(",", returnlist);
 
     }
 
     private String AddHoliday(List<String> Addholiday,List<String> holiday) {
+
+
+
         List<String> newHoliday= new ArrayList<>();
+
         if(!holiday.get(0).equals("")) newHoliday.addAll(holiday);
         for(String add: Addholiday) {
                 boolean redundancy=false;
@@ -120,30 +225,13 @@ public class DesignerHolidayService {
                 }
                 if(!redundancy) newHoliday.add(add);
         }
-        return String.join(",", newHoliday.toArray(new String[0]));
 
-    }
-    private DesignerHoliday findHoliday() {
-        long currentId= SecurityUtil.getCurrentMemberId();
-        User user=userRepository.findById(currentId)
-                .orElseThrow(() -> new RuntimeException("로그인 상태가 아닙니다"));
-        if(user.getRole() != Role.ROLE_DESIGNER) throw new RuntimeException("헤어디자이너만 접근 가능합니다");
-        return designerHolidayRepository.findBydesignerId(currentId);
-    }
 
-    public void delete() throws IOException {
-        DesignerHoliday holiday =findHoliday();
-        try {
-            if(holiday==null) throw new RuntimeException();
-            else {
 
-                designerHolidayRepository.delete(holiday);
-            }
 
-        }
-        catch(Exception e) {
-            throw new RuntimeException("휴일 삭제를 실패했습니다");
-        }
+
+        String[] returnlist= newHoliday.toArray(new String[0]);
+        return String.join(",", returnlist);
 
     }
 }
